# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": { "title": "API Documentation", "version": "1.0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "Bearer Authentication": [] }],
...
  "paths": {
    "/api/zones/{id}": {
      "put": {
        "tags": ["Zone Management"],
        "summary": "Update a zone",
        "description": "Update a zone by its ID.",
        "operationId": "updateZone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ZoneRequestDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseZoneDTO" }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouseId}": {
      "put": {
        "tags": ["Warehouse Management"],
        "summary": "Update a warehouse by ID",
        "description": "Update the details of an existing warehouse by ID",
        "operationId": "updateWarehouse",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WarehouseUpdateDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseWarehouse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shops/{id}": {
      "get": {
        "tags": ["Shop Management"],
        "summary": "Get shop by id",
        "description": "Fetch shop details by shop id. Accessible only if the current user is an admin, staff or owner of the shop.",
        "operationId": "getShopById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseShopDTO" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Shop Management"],
        "summary": "Update a shop",
        "description": "Update a shop by its ID.",
        "operationId": "updateShop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateShopDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseShopDTO" }
              }
            }
          }
        }
      }
    },
...
    "/api/zones": {
      "get": {
        "tags": ["Zone Management"],
        "summary": "Get all zones",
        "description": "Fetch a list of all registered zones.",
        "operationId": "getAllZonesOfUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListZoneDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Zone Management"],
        "summary": "Create a zone",
        "description": "Create a zone by owner or staff of the shop.",
        "operationId": "createZone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ZoneRequestDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseZoneDTO" }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{shopId}": {
      "post": {
        "tags": ["Warehouse Management"],
        "summary": "Create a new warehouse for a shop",
        "description": "Create a new warehouse under the specified shop",
        "operationId": "createWarehouse",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WarehouseCreateDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseWarehouse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/zones/{warehouseId}": {
      "get": {
        "tags": ["Zone Management"],
        "summary": "Get zones in warehouse",
        "description": "Fetch a list of zones in a warehouse of a shop.",
        "operationId": "getZonesByWarehouseId",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListZoneDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses": {
      "get": {
        "tags": ["Warehouse Management"],
        "summary": "Get paginate warehouses of a shop",
        "description": "Get paginate warehouses of a shop",
        "operationId": "getWarehouses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePaginateResponseWarehouseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{id}": {
      "get": {
        "tags": ["Warehouse Management"],
        "summary": "Get details warehouse of a shop",
        "description": "Get details warehouse of a shop",
        "operationId": "getWarehouseDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseWarehouseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/all": {
      "get": {
        "tags": ["Warehouse Management"],
        "summary": "Get all warehouses of a shop",
        "description": "Get all warehouses of a shop",
        "operationId": "getAllWarehouses",
        "parameters": [
          {
            "name": "shopId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListWarehouseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shops": {
      "get": {
        "tags": ["Shop Management"],
        "summary": "Get all shops",
        "description": "Fetch a list of all registered shops.",
        "operationId": "getShops",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePaginateResponseShopDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ZoneRequestDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "warehouseId": { "type": "integer", "format": "int64" }
        },
        "required": ["warehouseId"]
      },
      "BaseResponseZoneDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ZoneDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "ZoneDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "warehouseId": { "type": "integer", "format": "int64" },
          "warehouseName": { "type": "string" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
        }
      },
      "WarehouseUpdateDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "address": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "BaseResponseWarehouse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/Warehouse" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "Warehouse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "address": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
        },
        "required": ["status"]
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "role": { "type": "string" },
          "status": { "type": "string" },
          "shopId": { "type": "integer", "format": "int64" },
          "shopName": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": [
          "createdAt",
          "email",
          "id",
          "name",
          "phoneNumber",
          "role",
          "status",
          "username"
        ]
      },
      "UpdateShopDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "address": { "type": "string" }
        }
      },
      "BaseResponseShopDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ShopDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "ShopDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDTO" }
          },
          "createdBy": { "$ref": "#/components/schemas/UserDTO" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["createdAt"]
      },
...
      "WarehouseCreateDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "address": { "type": "string" },
          "description": { "type": "string" },
          "shopId": { "type": "integer", "format": "int64" }
        },
        "required": ["shopId"]
      },
...
      "BaseResponseListZoneDTO": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ZoneDTO" }
          },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "BaseResponsePaginateResponseWarehouseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginateResponseWarehouseDTO"
          },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "PaginateResponseWarehouseDTO": {
        "type": "object",
        "properties": {
          "pageSize": { "type": "integer", "format": "int32" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int32" },
          "totalPages": { "type": "integer", "format": "int32" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WarehouseDTO" }
          }
        }
      },
      "WarehouseDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "address": { "type": "string" },
          "shopId": { "type": "integer", "format": "int64" },
          "id": { "type": "integer", "format": "int64" },
          "shopName": { "type": "string" },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "numberOfZone": { "type": "integer", "format": "int32" }
        },
        "required": ["status"]
      },
      "BaseResponseWarehouseDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/WarehouseDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "BaseResponseListWarehouseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WarehouseDTO" }
          },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
...
      "BaseResponsePaginateResponseShopDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PaginateResponseShopDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "PaginateResponseShopDTO": {
        "type": "object",
        "properties": {
          "pageSize": { "type": "integer", "format": "int32" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int32" },
          "totalPages": { "type": "integer", "format": "int32" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ShopDTO" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}